[Refactor] monster-race-hook.cpp/h をpit-nest-util.cpp/h に吸収合併した
