+++ src/core/player-processor.cpp
@@ -8,7 +8,6 @@
 #include "core/window-redrawer.h"
 #include "floor/floor-save-util.h"
 #include "floor/floor-util.h"
-#include "floor/geometry.h"
 #include "floor/wild.h"
 #include "game-option/cheat-options.h"
 #include "game-option/disturbance-options.h"
@@ -23,12 +22,9 @@
 #include "mind/mind-sniper.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-describer.h"
-#include "monster/monster-flag-types.h"
 #include "monster/monster-list.h"
 #include "monster/monster-status-setter.h"
-#include "monster/monster-status.h"
 #include "monster/monster-update.h"
 #include "monster/monster-util.h"
 #include "mutation/mutation-investor-remover.h"
@@ -41,25 +37,18 @@
 #include "player/attack-defense-types.h"
 #include "player/eldritch-horror.h"
 #include "player/player-skill.h"
-#include "player/special-defense-types.h"
 #include "spell-kind/spells-random.h"
 #include "spell-realm/spells-hex.h"
 #include "spell-realm/spells-song.h"
 #include "status/action-setter.h"
-#include "system/angband-system.h"
-#include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
-#include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monrace/monrace-list.h"
-#include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "term/screen-processor.h"
 #include "timed-effect/timed-effects.h"
 #include "tracking/health-bar-tracker.h"
-#include "util/bit-flags-calculator.h"
 #include "view/display-messages.h"
-#include "window/display-sub-windows.h"
 #include "world/world-turn-processor.h"
 
 bool load = true;
