+++ src/effect/effect-monster-charm.cpp
@@ -5,26 +5,18 @@
 #include "effect/spells-effect-util.h"
 #include "monster-floor/monster-remover.h"
 #include "monster-race/monster-kind-mask.h"
-#include "monster-race/monster-race-hook.h"
-#include "monster/monster-flag-types.h"
 #include "monster/monster-info.h"
-#include "monster/monster-list.h"
 #include "monster/monster-status-setter.h"
-#include "monster/monster-status.h"
-#include "object-enchant/trc-types.h"
 #include "pet/pet-fall-off.h"
 #include "pet/pet-util.h"
 #include "player-base/player-class.h"
 #include "player/player-status-flags.h"
 #include "spell/spells-diceroll.h"
 #include "status/bad-status-setter.h"
-#include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
-#include "system/player-type-definition.h"
-#include "util/bit-flags-calculator.h"
 #include "view/display-messages.h"
 
 static void effect_monster_charm_resist(PlayerType *player_ptr, EffectMonster *em_ptr)
