+++ src/effect/effect-monster-switcher.cpp
@@ -9,7 +9,6 @@
 #include "effect/effect-monster-switcher.h"
 #include "avatar/avatar.h"
 #include "cmd-action/cmd-attack.h"
-#include "effect/attribute-types.h"
 #include "effect/effect-monster-charm.h"
 #include "effect/effect-monster-curse.h"
 #include "effect/effect-monster-evil.h"
@@ -20,19 +19,13 @@
 #include "effect/effect-monster-spirit.h"
 #include "effect/effect-monster-util.h"
 #include "mind/mind-elementalist.h"
-#include "monster-floor/monster-death.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-info.h"
 #include "monster/monster-status-setter.h"
-#include "monster/monster-status.h"
-#include "monster/monster-util.h"
-#include "player/player-damage.h"
 #include "spell-kind/spells-genocide.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
 #include "system/player-type-definition.h"
-#include "util/bit-flags-calculator.h"
 #include "view/display-messages.h"
 
 ProcessResult effect_monster_hypodynamia(PlayerType *player_ptr, EffectMonster *em_ptr)
