+++ src/melee/melee-postprocess.cpp
@@ -12,36 +12,22 @@
 
 #include "melee/melee-postprocess.h"
 #include "core/disturbance.h"
-#include "effect/attribute-types.h"
-#include "floor/cave.h"
 #include "floor/geometry.h"
-#include "grid/grid.h"
 #include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
-#include "monster-attack/monster-attack-table.h"
 #include "monster-floor/monster-death.h"
-#include "monster-floor/monster-move.h"
 #include "monster-floor/monster-remover.h"
-#include "monster-race/monster-race-hook.h"
-#include "monster-race/race-flags-resistance.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-description-types.h"
-#include "monster/monster-info.h"
 #include "monster/monster-status-setter.h"
 #include "monster/monster-status.h"
 #include "pet/pet-fall-off.h"
-#include "player-info/class-info.h"
-#include "player-info/race-types.h"
-#include "player/player-personality-types.h"
-#include "system/angband-system.h"
 #include "system/floor/floor-info.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "tracking/health-bar-tracker.h"
-#include "util/bit-flags-calculator.h"
-#include "util/string-processor.h"
 #include "view/display-messages.h"
 #include <string>
