+++ src/mind/mind-samurai.cpp
@@ -12,14 +12,10 @@
 #include "io/input-key-acceptor.h"
 #include "mind/stances-table.h"
 #include "monster-attack/monster-attack-player.h"
-#include "monster-race/monster-race-hook.h"
-#include "monster-race/race-flags-resistance.h"
 #include "monster-race/race-resistance-mask.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-info.h"
 #include "monster/monster-status-setter.h"
-#include "monster/monster-status.h"
-#include "object-enchant/tr-types.h"
 #include "pet/pet-util.h"
 #include "player-attack/player-attack.h"
 #include "player-base/player-class.h"
@@ -30,12 +26,9 @@
 #include "system/item-entity.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
-#include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "term/screen-processor.h"
-#include "term/z-form.h"
 #include "timed-effect/timed-effects.h"
-#include "util/bit-flags-calculator.h"
 #include "util/int-char-converter.h"
 #include "view/display-messages.h"
