+++ src/monster-floor/monster-death.cpp
@@ -5,7 +5,6 @@
 #include "effect/effect-characteristics.h"
 #include "effect/effect-processor.h"
 #include "floor/floor-object.h"
-#include "game-option/birth-options.h"
 #include "game-option/play-record-options.h"
 #include "io/write-diary.h"
 #include "main/music-definitions-table.h"
@@ -15,24 +14,16 @@
 #include "monster-floor/monster-death-util.h"
 #include "monster-floor/monster-object.h"
 #include "monster-floor/special-death-switcher.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster-race/race-brightness-mask.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-description-types.h"
-#include "monster/monster-flag-types.h"
-#include "monster/monster-info.h"
-#include "monster/monster-list.h"
 #include "object-enchant/item-apply-magic.h"
 #include "object-enchant/item-magic-applier.h"
 #include "pet/pet-fall-off.h"
 #include "player/patron.h"
 #include "sv-definition/sv-other-types.h"
 #include "sv-definition/sv-scroll-types.h"
-#include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
-#include "system/baseitem/baseitem-definition.h"
-#include "system/baseitem/baseitem-list.h"
-#include "system/building-type-definition.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
@@ -43,12 +34,9 @@
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "system/services/baseitem-monrace-service.h"
-#include "system/system-variables.h"
 #include "timed-effect/timed-effects.h"
-#include "util/bit-flags-calculator.h"
 #include "view/display-messages.h"
 #include "world/world.h"
-#include <algorithm>
 
 static void write_pet_death(PlayerType *player_ptr, MonsterDeath *md_ptr)
 {
