+++ src/monster-floor/monster-summon.cpp
@@ -1,20 +1,14 @@
 #include "monster-floor/monster-summon.h"
-#include "dungeon/dungeon-flag-types.h"
 #include "floor/geometry.h"
 #include "floor/wild.h"
 #include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/monster-race-hook.h"
-#include "monster/monster-info.h"
 #include "monster/monster-list.h"
 #include "monster/monster-util.h"
 #include "spell/summon-types.h"
-#include "system/dungeon/dungeon-definition.h"
-#include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
-#include "system/monrace/monrace-definition.h"
 #include "system/monrace/monrace-list.h"
 #include "system/monster-entity.h"
 #include "system/player-type-definition.h"
