+++ src/player-attack/player-attack.cpp
@@ -10,27 +10,18 @@
 #include "cmd-action/cmd-attack.h"
 #include "combat/attack-accuracy.h"
 #include "combat/attack-criticality.h"
-#include "combat/martial-arts-table.h"
 #include "combat/slaying.h"
-#include "floor/cave.h"
 #include "floor/geometry.h"
 #include "game-option/cheat-types.h"
-#include "inventory/inventory-slot-types.h"
-#include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
 #include "mind/mind-ninja.h"
 #include "mind/mind-samurai.h"
 #include "mind/monk-attack.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-damage.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-status-setter.h"
 #include "monster/monster-status.h"
-#include "monster/monster-util.h"
-#include "object-enchant/tr-types.h"
 #include "object-enchant/vorpal-weapon.h"
-#include "object-hook/hook-weapon.h"
-#include "object/tval-types.h"
 #include "player-attack/attack-chaos-effect.h"
 #include "player-attack/blood-sucking-processor.h"
 #include "player-base/player-class.h"
@@ -40,7 +31,6 @@
 #include "player/player-damage.h"
 #include "player/player-skill.h"
 #include "player/player-status-flags.h"
-#include "realm/realm-hex-numbers.h"
 #include "spell-kind/earthquake.h"
 #include "spell-realm/spells-hex.h"
 #include "sv-definition/sv-weapon-types.h"
@@ -50,13 +40,10 @@
 #include "system/item-entity.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
-#include "system/player-type-definition.h"
 #include "timed-effect/timed-effects.h"
-#include "util/bit-flags-calculator.h"
 #include "util/string-processor.h"
 #include "view/display-messages.h"
 #include "wizard/wizard-messages.h"
-#include "world/world.h"
 
 /*! 吸血処理の最大回復HP */
 constexpr auto MAX_VAMPIRIC_DRAIN = 50;
