+++ src/player/patron.cpp
@@ -1,6 +1,4 @@
 #include "player/patron.h"
-#include "cmd-action/cmd-pet.h"
-#include "cmd-io/cmd-dump.h"
 #include "flavor/flavor-describer.h"
 #include "flavor/object-flavor-types.h"
 #include "inventory/inventory-slot-types.h"
@@ -8,15 +6,10 @@
 #include "mind/mind-chaos-warrior.h"
 #include "monster-floor/monster-summon.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/monster-race-hook.h"
-#include "mutation/mutation-flag-types.h"
 #include "mutation/mutation-investor-remover.h"
-#include "object-enchant/object-curse.h"
 #include "player-base/player-class.h"
 #include "player-base/player-race.h"
-#include "player-info/class-info.h"
 #include "player-info/equipment-info.h"
-#include "player-info/race-types.h"
 #include "player/player-damage.h"
 #include "spell-kind/spells-floor.h"
 #include "spell-kind/spells-genocide.h"
@@ -32,7 +25,6 @@
 #include "status/shape-changer.h"
 #include "system/floor/floor-info.h"
 #include "system/item-entity.h"
-#include "system/player-type-definition.h"
 #include "view/display-messages.h"
 
 std::vector<Patron> patron_list = {
