+++ src/realm/realm-hissatsu.cpp
@@ -1,13 +1,9 @@
 #include "realm/realm-hissatsu.h"
 #include "artifact/fixed-art-types.h"
 #include "cmd-action/cmd-attack.h"
-#include "cmd-action/cmd-spell.h"
 #include "cmd-item/cmd-throw.h"
-#include "combat/combat-options-type.h"
 #include "core/asking-player.h"
 #include "core/stuff-handler.h"
-#include "dungeon/dungeon-flag-types.h"
-#include "effect/attribute-types.h"
 #include "effect/effect-characteristics.h"
 #include "effect/effect-processor.h"
 #include "effect/spells-effect-util.h"
@@ -18,12 +14,10 @@
 #include "io/input-key-acceptor.h"
 #include "io/input-key-requester.h"
 #include "mind/mind-ninja.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster-race/race-brightness-mask.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-info.h"
 #include "monster/monster-update.h"
-#include "object-enchant/tr-types.h"
 #include "player-info/equipment-info.h"
 #include "player/player-damage.h"
 #include "player/player-move.h"
@@ -50,7 +44,6 @@
 #include "target/target-getter.h"
 #include "term/screen-processor.h"
 #include "timed-effect/timed-effects.h"
-#include "util/bit-flags-calculator.h"
 #include "view/display-messages.h"
 #include "world/world.h"
