+++ src/room/pit-nest-util.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "monster-race/race-ability-flags.h"
+#include "util/flag-group.h"
 #include <map>
 #include <optional>
 #include <string>
@@ -51,6 +53,38 @@ struct nest_pit_type {
     int chance; //!< 生成確率
 
     void prepare_filter(PlayerType *player_ptr) const;
+};
+
+enum class MonraceId : short;
+enum class NestKind;
+enum class PitKind;
+class PitNestFilter {
+public:
+    ~PitNestFilter() = default;
+    PitNestFilter(const PitNestFilter &) = delete;
+    PitNestFilter(PitNestFilter &&) = delete;
+    PitNestFilter &operator=(const PitNestFilter &) = delete;
+    PitNestFilter &operator=(PitNestFilter &&) = delete;
+    static PitNestFilter &get_instance();
+
+    MonraceId get_monrace_id() const;
+    char get_monrace_symbol() const;
+    const EnumClassFlagGroup<MonsterAbilityType> &get_dragon_breaths() const;
+    std::string pit_subtype(PitKind type) const;
+    std::string nest_subtype(NestKind type) const;
+
+    void set_monrace_id(MonraceId id);
+    void set_monrace_symbol(char symbol);
+    void set_dragon_breaths();
+
+private:
+    PitNestFilter() = default;
+
+    static PitNestFilter instance;
+
+    MonraceId monrace_id{}; //<! 通常pit/nest生成時のモンスターの構成条件ID.
+    char monrace_symbol = '\0'; //!< 単一シンボルpit/nest生成時の指定シンボル.
+    EnumClassFlagGroup<MonsterAbilityType> dragon_breaths{}; //!< ブレス属性に基づくドラゴンpit生成時条件マスク.
 };
 
 /*! デバッグ時にnestのモンスター情報を確認するための構造体 / A struct for nest monster information with cheat_hear */
