+++ src/room/rooms-pit.cpp
@@ -5,9 +5,7 @@
 #include "grid/grid.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-util.h"
-#include "room/door-definition.h"
 #include "room/pit-nest-util.h"
 #include "room/space-finder.h"
 #include "system/enums/dungeon/dungeon-id.h"
