+++ src/spell-kind/spells-detection.cpp
@@ -1,19 +1,10 @@
 #include "spell-kind/spells-detection.h"
-#include "core/window-redrawer.h"
-#include "dungeon/dungeon-flag-types.h"
-#include "floor/cave.h"
 #include "floor/floor-save-util.h"
 #include "grid/grid.h"
 #include "grid/trap.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-flag-types.h"
-#include "monster/monster-info.h"
-#include "monster/monster-status.h"
 #include "monster/monster-update.h"
-#include "object/object-mark-types.h"
-#include "object/tval-types.h"
 #include "realm/realm-song-numbers.h"
-#include "realm/realm-song.h"
 #include "spell-realm/spells-song.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
