+++ src/system/building-type-definition.cpp
@@ -1,11 +1,9 @@
 #include "system/building-type-definition.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/monster-race-hook.h"
 #include "monster/monster-list.h"
 #include "monster/monster-util.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monrace/monrace-list.h"
-#include "system/player-type-definition.h"
 #include "system/services/dungeon-service.h"
 #include <numeric>
